move id replacement into the compiler lib
