+++ UnitTests/UnitTests.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TestsCommon\TestsCommon.csproj" />
+    <ProjectReference Include="..\msgraph-sdk-raptor-compiler-lib\msgraph-sdk-raptor-compiler-lib.csproj" />
   </ItemGroup>
 
   <ItemGroup>
