+++ msgraph-sdk-raptor.sln
@@ -34,7 +34,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CsharpBetaExecutionTests", "CsharpBetaExecutionTests\CsharpBetaExecutionTests.csproj", "{295FB66C-5A96-4FDE-BB01-38E3DC91AD87}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestsCommon.Tests", "TestsCommon.Tests\TestsCommon.Tests.csproj", "{28AA9EBB-3F91-4E63-B0F8-2BB165927B3D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "UnitTests\UnitTests.csproj", "{28AA9EBB-3F91-4E63-B0F8-2BB165927B3D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
